tenstorrent: bh_arc: remove init.c
Let's remove init.c and ensure that Zephyr's init system is doing the
heavy lifting.
